test: reorganize tests utils
