--- tests/helpers/util.lua
@@ -1,11 +1,0 @@
-local utils = {}
-
-utils.editfile = function(testfile)
-  vim.cmd("e " .. testfile)
-  assert.are.same(
-    vim.fn.fnamemodify(vim.api.nvim_buf_get_name(0), ":p"),
-    vim.fn.fnamemodify(testfile, ":p")
-  )
-end
-
-return utils
